Extract and move SPIs
