+++ src/main/java/org/opentripplanner/graph_builder/module/osm/OSMDatabase.java
@@ -50,7 +50,7 @@ import org.opentripplanner.street.search.TraverseModeSet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class OSMDatabase {
+public class OSMDatabase implements org.opentripplanner.openstreetmap.spi.OSMDatabase {
 
   private static final Logger LOG = LoggerFactory.getLogger(OSMDatabase.class);
 
@@ -209,6 +209,7 @@ public class OSMDatabase {
     return waysNodeIds.contains(nodeId);
   }
 
+  @Override
   public void addNode(OSMNode node) {
     if (node.isBikeParking()) {
       bikeParkingNodes.put(node.getId(), node);
@@ -232,6 +233,7 @@ public class OSMDatabase {
     nodesById.put(node.getId(), node);
   }
 
+  @Override
   public void addWay(OSMWay way) {
     /* only add ways once */
     long wayId = way.getId();
@@ -286,6 +288,7 @@ public class OSMDatabase {
     waysById.put(wayId, way);
   }
 
+  @Override
   public void addRelation(OSMRelation relation) {
     if (relationsById.containsKey(relation.getId())) {
       return;
@@ -329,16 +332,12 @@ public class OSMDatabase {
     relationsById.put(relation.getId(), relation);
   }
 
-  /**
-   * Called after the first phase, when all relations are loaded.
-   */
+  @Override
   public void doneFirstPhaseRelations() {
     // nothing to do here
   }
 
-  /**
-   * Called after the second phase, when all ways are loaded.
-   */
+  @Override
   public void doneSecondPhaseWays() {
     // This copies relevant tags to the ways (highway=*) where it doesn't exist, so that
     // the way purging keeps the needed way around.
@@ -354,10 +353,7 @@ public class OSMDatabase {
     markNodesForKeeping(areaWaysById.valueCollection(), areaNodeIds);
   }
 
-  /**
-   * Called after the third and final phase, when all nodes are loaded. After all relations, ways,
-   * and nodes are loaded, handle areas.
-   */
+  @Override
   public void doneThirdPhaseNodes() {
     processMultipolygonRelations();
     processSingleWayAreas();
