+++ src/main/java/org/opentripplanner/openstreetmap/OpenStreetMapParser.java
@@ -6,14 +6,14 @@ import java.util.Map;
 import java.util.Objects;
 import org.openstreetmap.osmosis.osmbinary.BinaryParser;
 import org.openstreetmap.osmosis.osmbinary.Osmformat;
-import org.opentripplanner.graph_builder.module.osm.OSMDatabase;
 import org.opentripplanner.openstreetmap.model.OSMNode;
 import org.opentripplanner.openstreetmap.model.OSMNodeRef;
-import org.opentripplanner.openstreetmap.model.OSMProvider;
 import org.opentripplanner.openstreetmap.model.OSMRelation;
 import org.opentripplanner.openstreetmap.model.OSMRelationMember;
 import org.opentripplanner.openstreetmap.model.OSMTag;
 import org.opentripplanner.openstreetmap.model.OSMWay;
+import org.opentripplanner.openstreetmap.spi.OSMDatabase;
+import org.opentripplanner.openstreetmap.spi.OSMProvider;
 
 /**
  * Parser for the OpenStreetMap PBF Format.
