+++ src/main/java/org/opentripplanner/openstreetmap/OpenStreetMapProvider.java
@@ -13,12 +13,12 @@ import org.opentripplanner.framework.application.OtpFileNames;
 import org.opentripplanner.framework.logging.ProgressTracker;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
 import org.opentripplanner.graph_builder.ConfiguredDataSource;
-import org.opentripplanner.graph_builder.module.osm.OSMDatabase;
 import org.opentripplanner.graph_builder.module.osm.WayPropertySet;
 import org.opentripplanner.graph_builder.module.osm.parameters.OsmExtractParameters;
 import org.opentripplanner.graph_builder.module.osm.parameters.OsmExtractParametersBuilder;
 import org.opentripplanner.graph_builder.module.osm.tagmapping.OsmTagMapper;
-import org.opentripplanner.openstreetmap.model.OSMProvider;
+import org.opentripplanner.openstreetmap.spi.OSMDatabase;
+import org.opentripplanner.openstreetmap.spi.OSMProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
