+++ src/main/java/org/opentripplanner/openstreetmap/model/OSMWithTags.java
@@ -10,6 +10,7 @@ import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
 import org.opentripplanner.openstreetmap.TemplateLibrary;
+import org.opentripplanner.openstreetmap.spi.OSMProvider;
 import org.opentripplanner.transit.model.basic.Accessibility;
 import org.opentripplanner.transit.model.basic.I18NString;
 import org.opentripplanner.transit.model.basic.NonLocalizedString;
